requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.3
beautifulsoup4==4.5.3
certifi==2017.7.27.1
cffi==1.11.1
click==6.7
colorama==0.3.9
coverage==4.5.1
cycler==0.10.0
Django==1.11
django-bootstrap3==9.1.0
django-braces==1.11.0
django-extensions==2.0.7
django-model-utils==3.0.0
django-star-ratings==0.5.6
djangorestframework==3.7.1
Faker==0.8.8
Flask==0.12.1
h5py==2.7.1
itsdangerous==0.24
Jinja2==2.9.6
Keras==2.1.6
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
mixer==6.0.1
more-itertools==4.3.0
numpy==1.13.3
olefile==0.44
pandas==0.21.0
pathlib2==2.3.2
Pillow==4.3.0
pluggy==0.7.1
protobuf==3.1.0
py==1.6.0
pycparser==2.18
pydot==1.2.4
PyMySQL==0.7.11
pyparsing==2.2.0
pytest==3.8.0
pytest-cov==2.6.0
pytest-django==3.4.2
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.13.0
scipy==1.1.0
six==1.11.0
tensorflow==0.12.0
text-unidecode==1.2
Theano==1.0.1+100.g2c19431b1
urllib3==1.20
virtualenv==15.1.0
virtualenvwrapper-win==1.2.2
Werkzeug==0.12.1
